calculator-service tax response

+++ application-service/src/main/java/calculatorResponse/CalculatorResponse.java
@@ -1,4 +1,4 @@
 package calculatorResponse;
 
-public record CalculatorResponse(int tax) {
+public record CalculatorResponse(double tax) {
 }
